+++ internal/component/otelcol/config_http.go
@@ -21,13 +21,9 @@ type HTTPServerArguments struct {
 
 	CORS *CORSArguments `alloy:"cors,block,optional"`
 
-	// TODO(rfratto): auth
-	//
-	// Figuring out how to do authentication isn't very straightforward here. The
-	// auth section links to an authenticator extension.
-	//
-	// We will need to generally figure out how we want to provide common
-	// authentication extensions to all of our components.
+	// Auth is a binding to an otelcol.auth.* component extension which handles
+	// authentication.
+	Auth *auth.Handler `alloy:"auth,attr,optional"`
 
 	MaxRequestBodySize    units.Base2Bytes `alloy:"max_request_body_size,attr,optional"`
 	IncludeMetadata       bool             `alloy:"include_metadata,attr,optional"`
@@ -44,9 +40,25 @@ func copyStringSlice(s []string) []string {
 }
 
 // Convert converts args into the upstream type.
-func (args *HTTPServerArguments) Convert() *otelconfighttp.ServerConfig {
+func (args *HTTPServerArguments) Convert() (*otelconfighttp.ServerConfig, error) {
 	if args == nil {
-		return nil
+		return nil, nil
+	}
+
+	// If auth is set by the user retrieve the associated extension from the handler.
+	// if the extension does not support server auth an error will be returned.
+	var authz *otelconfighttp.AuthConfig
+	if args.Auth != nil {
+		ext, err := args.Auth.GetExtension(auth.Server)
+		if err != nil {
+			return nil, err
+		}
+
+		authz = &otelconfighttp.AuthConfig{
+			Authentication: otelconfigauth.Authentication{
+				AuthenticatorID: ext.ID,
+			},
+		}
 	}
 
 	return &otelconfighttp.ServerConfig{
@@ -56,7 +68,22 @@ func (args *HTTPServerArguments) Convert() *otelconfighttp.ServerConfig {
 		MaxRequestBodySize:    int64(args.MaxRequestBodySize),
 		IncludeMetadata:       args.IncludeMetadata,
 		CompressionAlgorithms: copyStringSlice(args.CompressionAlgorithms),
+		Auth:                  authz,
+	}, nil
+}
+
+// Extensions exposes extensions used by args.
+func (args *HTTPServerArguments) Extensions() map[otelcomponent.ID]otelextension.Extension {
+	m := make(map[otelcomponent.ID]otelextension.Extension)
+	if args.Auth != nil {
+		ext, err := args.Auth.GetExtension(auth.Server)
+		// Extension will not be registered if there was an error.
+		if err != nil {
+			return m
+		}
+		m[ext.ID] = ext.Extension
 	}
+	return m
 }
 
 // CORSArguments holds shared CORS settings for components which launch HTTP
@@ -113,15 +140,19 @@ type HTTPClientArguments struct {
 }
 
 // Convert converts args into the upstream type.
-func (args *HTTPClientArguments) Convert() *otelconfighttp.ClientConfig {
+func (args *HTTPClientArguments) Convert() (*otelconfighttp.ClientConfig, error) {
 	if args == nil {
-		return nil
+		return nil, nil
 	}
 
 	// Configure the authentication if args.Auth is set.
-	var auth *otelconfigauth.Authentication
+	var authz *otelconfigauth.Authentication
 	if args.Auth != nil {
-		auth = &otelconfigauth.Authentication{AuthenticatorID: args.Auth.ID}
+		ext, err := args.Auth.GetExtension(auth.Client)
+		if err != nil {
+			return nil, err
+		}
+		authz = &otelconfigauth.Authentication{AuthenticatorID: ext.ID}
 	}
 
 	opaqueHeaders := make(map[string]configopaque.String)
@@ -150,17 +181,21 @@ func (args *HTTPClientArguments) Convert() *otelconfighttp.ClientConfig {
 		HTTP2ReadIdleTimeout: args.HTTP2ReadIdleTimeout,
 		HTTP2PingTimeout:     args.HTTP2PingTimeout,
 
-		Auth: auth,
+		Auth: authz,
 
 		Cookies: args.Cookies.Convert(),
-	}
+	}, nil
 }
 
 // Extensions exposes extensions used by args.
 func (args *HTTPClientArguments) Extensions() map[otelcomponent.ID]otelextension.Extension {
 	m := make(map[otelcomponent.ID]otelextension.Extension)
 	if args.Auth != nil {
-		m[args.Auth.ID] = args.Auth.Extension
+		ext, err := args.Auth.GetExtension(auth.Client)
+		if err != nil {
+			return m
+		}
+		m[ext.ID] = ext.Extension
 	}
 	return m
 }